Allow both thread starter and last poster to be displayed on Mobile
Removed the "By: " and "Last: " labels as this information is
obvious based on the order of the names. Text overflows into
ellipsis allowing the number and timespan to be viewed while 'ago'
gets cut off in cases where there is not enough space for longer
names. E.g. '1 h...' or '1 y...'